+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.br.seventh_art">
 
-    <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -12,16 +11,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.7ART">
-        <activity android:name=".view.ProfileActivity"></activity>
-        <!--        <activity android:name=".view.SeriesCategoryContentActivity" />-->
+
+        <activity android:name=".view.ProfileActivity" /> <!-- <activity android:name=".view.genres.series.activity.SeriesGenresActivity" /> -->
         <activity android:name=".view.UserProfileActivity" />
         <activity android:name=".view.MainActivity" />
         <activity android:name=".view.MovieReadMoreActivity" />
-        <activity android:name=".view.CategoryContentActivity" />
         <activity android:name=".view.SignInActivity" />
         <activity android:name=".view.SignUpActivity" />
         <activity android:name=".view.LoginActivity" />
-        <activity android:name=".view.MovieActivity" />
+        <activity android:name=".view.genres.movies.activity.MoviesGenresActivity" />
         <activity android:name=".view.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
