+++ app/src/main/java/com/br/seventh_art/model/LoggedInUser.kt
@@ -1,0 +1,9 @@
+package com.br.seventh_art.model
+
+/**
+ * Data class that captures user information for logged in users retrieved from LoginRepository
+ */
+data class LoggedInUser(
+    val userId: String,
+    val displayName: String
+)
