+++ app/src/main/java/com/br/seventh_art/model/genreSerie/SerieGenreResponse.kt
@@ -1,4 +1,4 @@
-package com.br.seventh_art.data.model.genreSerie
+package com.br.seventh_art.model.genreSerie
 
 data class SerieGenreResponse(
     val genres: List<SerieGenre>
