+++ app/src/main/java/com/br/seventh_art/repository/Repository.kt
@@ -1,8 +1,8 @@
 package com.br.seventh_art.repository
 
 import RetrofitInit
-import com.br.seventh_art.data.model.genreMovie.MovieGenreResponse
-import com.br.seventh_art.data.model.genreSerie.SerieGenreResponse
+import com.br.seventh_art.model.genreMovie.MovieGenreResponse
+import com.br.seventh_art.model.genreSerie.SerieGenreResponse
 import com.br.seventh_art.network.EndPointApi
 
 const val API_KEY = "d00ab27062c01a80c4f6a7cefd66a6a5"
