+++ app/src/main/java/com/br/seventh_art/view/SignUpActivity.kt
@@ -8,7 +8,7 @@ import android.widget.ImageButton
 import androidx.appcompat.app.AppCompatActivity
 import com.br.seventh_art.R
 import com.br.seventh_art.ui.login.Utils
-import com.google.android.material.textfield.TextInputLayout
+import com.br.seventh_art.view.genres.MoviesGenresActivity
 
 class SignUpActivity : AppCompatActivity(), Utils {
 
@@ -29,7 +29,8 @@ class SignUpActivity : AppCompatActivity(), Utils {
     private fun initClick() {
         buttonLogIn.setOnClickListener{
             if(validateName(username) && validateEmail(emailSignUp) && validatePassword(passwordSignUp) && validateConfirmPassword(passwordSignUp,confirmPassword)){
-                val intent = Intent(this,CategoryContentActivity::class.java)
+                val intent = Intent(this,
+                    MoviesGenresActivity::class.java)
                 startActivity(intent)
             }
         }
