+++ app/src/main/res/values/themes.xml
@@ -23,4 +23,13 @@
 
     </style>
 
+    <style name="Theme.7ART.NoActionBar">
+        <item name="windowActionBar">false</item>
+        <item name="windowNoTitle">true</item>
+    </style>
+
+    <style name="Theme.7ART.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
+
+    <style name="Theme.7ART.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
+
 </resources>
